src/dex_client/dex_client.py
"""
DEX client for interacting with Solana DEX (Orca/Raydium).
"""
import os
from typing import Dict, Any, Optional
from solana.rpc.async_api import AsyncClient
from solana.keypair import Keypair
from solana.transaction import Transaction
import base58

class DexClient:
    def __init__(self):
        self.rpc_url = os.getenv("SOLANA_RPC_URL")
        self.private_key = os.getenv("PRIVATE_KEY")
        self.dex_target = os.getenv("DEX_TARGET", "orca").lower()
        
        # Initialize Solana client
        self.client = AsyncClient(self.rpc_url)
        
        # Load wallet
        self.wallet = self._load_wallet()
        
    def _load_wallet(self) -> Keypair:
        """Load wallet from private key."""
        if not self.private_key:
            raise ValueError("Private key not found in environment variables")
        
        try:
            private_key_bytes = base58.b58decode(self.private_key)
            return Keypair.from_secret_key(private_key_bytes)
        except Exception as e:
            raise ValueError(f"Failed to load wallet: {str(e)}")
    
    async def get_token_balance(self, token_address: str) -> float:
        """Get token balance for the wallet."""
        try:
            response = await self.client.get_token_account_balance(token_address)
            if response["result"]["value"]:
                return float(response["result"]["value"]["amount"]) / 10**response["result"]["value"]["decimals"]
            return 0.0
        except Exception as e:
            print(f"Error getting token balance: {str(e)}")
            return 0.0
    
    async def get_pool_info(self, pool_address: str) -> Dict[str, Any]:
        """Get pool information from DEX."""
        # This is a placeholder that should be implemented with actual DEX SDK
        return {
            "address": pool_address,
            "token_a": "SOL",
            "token_b": "USDC",
            "reserve_a": 0.0,
            "reserve_b": 0.0,
            "fee": 0.003  # 0.3%
        }
    
    async def create_order(self, side: str, price: float, size: float) -> Dict[str, Any]:
        """
        Create an order on the DEX.
        
        Args:
            side: "bid" or "ask"
            price: Order price
            size: Order size in USD
            
        Returns:
            Order information including order ID
        """
        # This is a placeholder that should be implemented with actual DEX SDK
        # The actual implementation would:
        # 1. Create the appropriate transaction
        # 2. Sign it with the wallet
        # 3. Send it to the network
        # 4. Return the order details
        
        print(f"Creating {side} order: {price} @ {size}")
        
        return {
            "id": f"order_{side}_{price}_{size}",
            "status": "pending",
            "side": side,
            "price": price,
            "size": size
        }
    
    async def cancel_order(self, order_id: str) -> bool:
        """Cancel an existing order."""
        # This is a placeholder that should be implemented with actual DEX SDK
        print(f"Cancelling order: {order_id}")
        return True
    
    async def get_order_status(self, order_id: str) -> Dict[str, Any]:
        """Get the status of an existing order."""
        # This is a placeholder that should be implemented with actual DEX SDK
        return {
            "id": order_id,
            "status": "open",
            "filled_amount": 0.0,
            "remaining_amount": 0.0
        }
    
    async def get_market_price(self, pool_address: str) -> Optional[float]:
        """Get current market price from pool."""
        try:
            pool_info = await self.get_pool_info(pool_address)
            if pool_info["reserve_a"] > 0 and pool_info["reserve_b"] > 0:
                return pool_info["reserve_b"] / pool_info["reserve_a"]
            return None
        except Exception as e:
            print(f"Error getting market price: {str(e)}")
            return None
    
    async def close(self):
        """Close the DEX client connection."""
        await self.client.close() 